+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 ## 1.0.0 (`1000070`)
 
 - **Fix:** Site key is not sent to hCaptcha when validating captcha (#3)
+- **Fix:** Enforcing hostname verification can cause captcha validation to fail (#4)
 
 ## 1.0.0 Alpha 1 (`1000011`)