+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 CHANGELOG
 ==========================
 
+## 1.0.0 (`1000070`)
+
+- **Fix:** Site key is not sent to hCaptcha when validating captcha (#3)
+- **Fix:** Enforcing hostname verification can cause captcha validation to fail (#4)
+- **Fix:** hCaptcha validation error codes are not logged (#5)
+
 ## 1.0.0 Alpha 1 (`1000011`)
 
 Initial release